+++ plu-api/src/main/kotlin/com/th/plu/api/service/redis/RedisHandler.kt
@@ -8,6 +8,10 @@ import java.util.concurrent.TimeUnit
 class RedisHandler(
     private val redisTemplate: RedisTemplate<String, Any>
 ) {
+    fun get(key: String): String? {
+        return redisTemplate.opsForValue().get(key) as String?
+    }
+
     fun set(key: String, value: Any, timeout: Long) {
         redisTemplate.opsForValue().set(key, value, timeout, TimeUnit.MILLISECONDS)
     }
